requirements_backend_bot.txt
fastapi
python-telegram-bot==13.15
requests
uvicorn
